+++ prisma/migrations/20260810164242_drop_campaign_card_opacity/migration.sql
@@ -1,0 +1,2 @@
+-- AlterTable
+ALTER TABLE "drop_campaigns" ADD COLUMN     "bg_card_opacity" SMALLINT NOT NULL DEFAULT 100;